% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision_matrix.R
\name{comp_lambda}
\alias{comp_lambda}
\title{Eigenvalues for Efficient Log-Determinant}
\usage{
comp_lambda(G)
}
\arguments{
\item{G}{The base precision matrix from \code{\link[=create_Q]{create_Q()}} with alpha = 1.}
}
\value{
A numeric vector of eigenvalues.
}
\description{
Computes the eigenvalues of the base Laplacian matrix G, which can be
reused to efficiently compute log|Q| for different values of kappa.
}
\details{
For the precision matrix Q = kappa^4\emph{I + 2}kappa^2*G + G'G, if G has
eigenvalues lambda_i, then Q has eigenvalues:
\deqn{\xi_i = \kappa^4 + 2\kappa^2 \lambda_i + \lambda_i^2}
and log|Q| = sum(log(xi_i)).

Computing the eigenvalues once and reusing them is much faster than
repeatedly computing log-determinants via Cholesky factorization.
}
\examples{
G <- create_Q(c(10, 10), alpha = 1)
lambda <- comp_lambda(G)
log_det_Q(kappa = 0.5, lambda)

}
\seealso{
\code{\link[=log_det_Q]{log_det_Q()}} which uses these eigenvalues.
}
