% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition.R
\name{compo_dist}
\alias{compo_dist}
\title{Aitchison Distance Between Compositions}
\usage{
compo_dist(x, y)
}
\arguments{
\item{x, y}{Compositional vectors or matrices. If matrices, they must
have the same dimensions, and the distance is computed row-wise.}
}
\value{
A numeric value (if vectors) or vector (if matrices) of
Aitchison distances.
}
\description{
Computes the Aitchison distance between two compositional vectors or
between rows of two compositional matrices.
}
\details{
The Aitchison distance is defined as the Euclidean distance between
CLR (centered log-ratio) transformed compositions:
\deqn{d_A(x, y) = \sqrt{\sum_{k=1}^{D} (\log(x_k/g(x)) - \log(y_k/g(y)))^2}}
where g(x) is the geometric mean of x.
}
\examples{
x <- c(0.3, 0.5, 0.2)
y <- c(0.4, 0.4, 0.2)
compo_dist(x, y)

}
