% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision_matrix.R
\name{create_W}
\alias{create_W}
\title{Create Second-Difference Matrix for 1D}
\usage{
create_W(n, neumann = TRUE)
}
\arguments{
\item{n}{Grid size.}

\item{neumann}{If TRUE, use Neumann boundaries; otherwise torus.}
}
\value{
Sparse matrix of dimension n x n.
}
\description{
Internal helper function to create the second-difference (Laplacian)
matrix for a 1D grid.
}
\keyword{internal}
