% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition.R
\name{dalr}
\alias{dalr}
\title{Jacobian of the Inverse ALR Transformation}
\usage{
dalr(z)
}
\arguments{
\item{z}{A matrix of compositional data with n rows (observations) and
D columns (components).}
}
\value{
A matrix of size (n*d) x D containing the Jacobian. The matrix
is structured in blocks where rows (k-1)\emph{n+1 to k}n correspond to
the derivatives of component k with respect to all D components.
}
\description{
Computes the Jacobian matrix of the inverse ALR transformation,
i.e., the partial derivatives dz/dx where z is compositional and
x is ALR-transformed.
}
\details{
The derivatives are:
\itemize{
\item \eqn{dz_k/dx_j = z_k(1 - z_k)} if k = j
\item \eqn{dz_k/dx_j = -z_k z_j} if k != j
}
}
\examples{
compo <- matrix(c(0.3, 0.5, 0.2), nrow = 1)
J <- dalr(compo)

}
\seealso{
\code{\link[=alr]{alr()}}, \code{\link[=invalr]{invalr()}}
}
