% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivatives.R
\name{fisher_info}
\alias{fisher_info}
\title{Fisher Information Matrix}
\usage{
fisher_info(x, alpha, A, y, b, c, Q, step, d, w)
}
\arguments{
\item{x}{Vector of ALR-transformed latent field values.}

\item{alpha}{Positive precision parameter.}

\item{A}{Location matrix.}

\item{y}{Observed compositional data.}

\item{b, c}{Hyperparameters for alpha prior.}

\item{Q}{GMRF precision matrix.}

\item{step}{Step size for MALA.}

\item{d}{Number of ALR components.}

\item{w}{Observation weights.}
}
\value{
A list with components:
\item{RFI}{Upper triangular Cholesky factor of the Fisher information}
\item{m}{Mean of the MALA proposal}
\item{p}{Permutation vector from AMD ordering}
}
\description{
Computes the Fisher information matrix for the MALA proposal, including
the Cholesky factor and mean of the Langevin proposal.
}
\details{
The Fisher information combines:
\itemize{
\item Negative Hessian of the Dirichlet log-likelihood
\item GMRF precision matrix Q
\item Second derivative w.r.t. alpha from the gamma prior
}

AMD (approximate minimum degree) ordering is used to reduce fill-in
during the Cholesky factorization.
}
\seealso{
\code{\link[=mala_step]{mala_step()}} which uses this function.
}
