% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs.R
\name{grad_log_posterior_kappa}
\alias{grad_log_posterior_kappa}
\title{Gradient of Log-Posterior of Kappa (w.r.t. log kappa)}
\usage{
grad_log_posterior_kappa(x, kappa, priors)
}
\arguments{
\item{x}{Matrix (N x d) of latent field values.}

\item{kappa}{Spatial range parameter (positive).}

\item{priors}{List containing prior hyperparameters.}
}
\value{
Gradient of log p(kappa | x) with respect to log(kappa).
}
\description{
Computes the gradient of the marginal log-posterior of kappa with respect
to log(kappa), for use in gradient-based kappa samplers.
}
\keyword{internal}
