% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivatives.R
\name{gradient_dirichlet}
\alias{gradient_dirichlet}
\title{Gradient of Dirichlet Log-Likelihood}
\usage{
gradient_dirichlet(z, alpha, w, y)
}
\arguments{
\item{z}{A matrix of compositional parameters (n x D) in (0,1).}

\item{alpha}{Positive precision parameter of the Dirichlet distribution.}

\item{w}{Observation weights, either scalar (uniform) or vector of length n.}

\item{y}{Observed compositional data (n x D) in (0,1).}
}
\value{
A list with components:
\item{g}{Vector of length n*d containing the gradient with respect to
the ALR-transformed field}
\item{dg}{Matrix (n x D) of partial derivatives with respect to z}
}
\description{
Computes the gradient of the Dirichlet log-likelihood with respect to
the ALR-transformed latent field.
}
\details{
The Dirichlet log-likelihood is:
\deqn{\log p(y|z,\alpha) = \sum_i [\log\Gamma(\alpha w_i) -
  \sum_k \log\Gamma(\alpha w_i z_{ik}) + \sum_k (\alpha w_i z_{ik} - 1)\log y_{ik}]}

The gradient with respect to z_k is:
\deqn{\frac{\partial}{\partial z_k} = -\alpha w \psi(\alpha w z_k) + \alpha w \log y_k}

The gradient with respect to the ALR-transformed field x is obtained by
chain rule using the Jacobian from \code{\link[=dalr]{dalr()}}.
}
\examples{
z <- matrix(c(0.3, 0.5, 0.2), nrow = 1)
y <- matrix(c(0.35, 0.45, 0.2), nrow = 1)
grad <- gradient_dirichlet(z, alpha = 10, w = 1, y)

}
\seealso{
\code{\link[=hessian_dirichlet]{hessian_dirichlet()}} for second derivatives, \code{\link[=neg_log_lik]{neg_log_lik()}} for
the full posterior.
}
