% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivatives.R
\name{hessian_dirichlet}
\alias{hessian_dirichlet}
\title{Hessian of Dirichlet Log-Likelihood}
\usage{
hessian_dirichlet(z, alpha, w, y)
}
\arguments{
\item{z}{A matrix of compositional parameters (n x D) in (0,1).}

\item{alpha}{Positive precision parameter of the Dirichlet distribution.}

\item{w}{Observation weights, either scalar or vector of length n.}

\item{y}{Observed compositional data (n x D) in (0,1).}
}
\value{
A list with components:
\item{H}{Sparse Hessian matrix of dimension (n\emph{d) x (n}d)}
\item{d2gii}{Matrix (n x d) of diagonal second derivatives}
\item{d2gij}{Matrix (n x d) of off-diagonal second derivatives}
}
\description{
Computes the Hessian matrix (second derivatives) of the Dirichlet
log-likelihood with respect to the ALR-transformed latent field.
}
\details{
The Hessian is computed using the chain rule:
\deqn{H = \frac{\partial^2 \log p}{\partial x \partial x^T} =
  \frac{\partial z}{\partial x}^T \frac{\partial^2 \log p}{\partial z^2}
  \frac{\partial z}{\partial x} + ...}

The resulting Hessian is sparse with a band structure determined by
the number of components.
}
\examples{
z <- matrix(c(0.3, 0.5, 0.2), nrow = 1)
y <- matrix(c(0.35, 0.45, 0.2), nrow = 1)
H <- hessian_dirichlet(z, alpha = 10, w = 1, y)

}
\seealso{
\code{\link[=gradient_dirichlet]{gradient_dirichlet()}} for first derivatives, \code{\link[=fisher_info]{fisher_info()}}
for the expected Fisher information.
}
