% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision_matrix.R
\name{log_det_Q}
\alias{log_det_Q}
\title{Log-Determinant of GMRF Precision Matrix}
\usage{
log_det_Q(kappa, lambda)
}
\arguments{
\item{kappa}{Positive spatial range parameter.}

\item{lambda}{Vector of eigenvalues from \code{\link[=comp_lambda]{comp_lambda()}}.}
}
\value{
The log-determinant log|Q|.
}
\description{
Efficiently computes the log-determinant of the GMRF precision matrix Q
using precomputed eigenvalues.
}
\details{
Uses the formula:
\deqn{\log|Q| = \sum_i \log(\kappa^4 + 2\kappa^2 \lambda_i + \lambda_i^2)}

This is O(N) once eigenvalues are computed, compared to \eqn{O(N^{1.5})} for
sparse Cholesky.
}
\examples{
G <- create_Q(c(10, 10), alpha = 1)
lambda <- comp_lambda(G)
log_det_Q(kappa = 0.5, lambda)
log_det_Q(kappa = 1.0, lambda)

}
\seealso{
\code{\link[=comp_lambda]{comp_lambda()}} for computing eigenvalues.
}
