% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{posterior_predictive}
\alias{posterior_predictive}
\title{Compute Posterior Predictive Samples}
\usage{
posterior_predictive(mcmc, n_samples = 100, burnin = NULL, A = NULL)
}
\arguments{
\item{mcmc}{Output from \code{\link[=mcmc_sampling]{mcmc_sampling()}}.}

\item{n_samples}{Number of posterior predictive samples. Default is 100.}

\item{burnin}{Burnin to discard. Default is iter/4.}

\item{A}{Location matrix (if different from identity).}
}
\value{
An array of dimension (N x D x n_samples) containing posterior
predictive compositions.
}
\description{
Generates posterior predictive samples of compositional observations.
}
\details{
For each posterior sample of (x, alpha), generates a new observation
from the Dirichlet likelihood. Useful for model checking by comparing
predictive distribution to observed data.
}
\examples{
\dontrun{
fit <- mcmc_sampling(theta0, priors, y, iter = 5000)
y_pred <- posterior_predictive(fit, n_samples = 50)

# Compare observed vs predicted
hist(y[, 1], main = "Component 1", freq = FALSE)
lines(density(y_pred[, 1, ]), col = "red")
}

}
