% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{rdirichlet}
\alias{rdirichlet}
\title{Random Dirichlet Distribution}
\usage{
rdirichlet(n, alpha)
}
\arguments{
\item{n}{Number of samples to generate.}

\item{alpha}{Vector of concentration parameters (all positive).}
}
\value{
A matrix with n rows and length(alpha) columns, where each row
is a sample from Dirichlet(alpha).
}
\description{
Generates random samples from the Dirichlet distribution.
}
\details{
The Dirichlet distribution is sampled by:
\enumerate{
\item Generate x_k ~ Gamma(alpha_k, 1) for each k
\item Normalize: y_k = x_k / sum(x)
}
}
\examples{
# Sample from Dirichlet(2, 3, 5)
samples <- spatiocompo:::rdirichlet(100, c(2, 3, 5))
colMeans(samples)  # Should be close to c(0.2, 0.3, 0.5)

}
\keyword{internal}
