% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{rinvwishart}
\alias{rinvwishart}
\title{Random Inverse Wishart Distribution}
\usage{
rinvwishart(nu, S)
}
\arguments{
\item{nu}{Degrees of freedom (must be > dim - 1).}

\item{S}{Scale matrix (positive definite).}
}
\value{
A positive definite matrix drawn from IW(nu, S).
The expected value is S / (nu - p - 1) where p is the dimension.
}
\description{
Generates a random sample from the inverse Wishart distribution.
}
\details{
Uses the Bartlett decomposition: sample \eqn{W \sim Wishart(\nu, S^{-1})}, return \eqn{W^{-1}}.
}
\examples{
# Sample from IW(10, I_2)
# E[X] = I / (10 - 2 - 1) = I / 7
set.seed(123)
samples <- replicate(1000, spatiocompo:::rinvwishart(10, diag(2)), simplify = FALSE)
mean_diag <- mean(sapply(samples, function(x) x[1,1]))
# Should be close to 1/7 = 0.143

}
\keyword{internal}
