% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs.R
\name{sample_invwishart}
\alias{sample_invwishart}
\title{Sample from Inverse-Wishart Conditional Posterior}
\usage{
sample_invwishart(x, kappa, priors)
}
\arguments{
\item{x}{Matrix (N x d) of latent field values.}

\item{kappa}{Spatial range parameter.}

\item{priors}{List containing prior hyperparameters:
\itemize{
\item field.G: Base precision matrix
\item rho.df: Prior degrees of freedom
\item rho.Sigma: Prior scale matrix
}}
}
\value{
A d x d positive definite covariance matrix.
}
\description{
Samples the covariance matrix rho from its conditional inverse-Wishart
posterior given the latent field x and spatial range kappa.
}
\details{
With the conjugate inverse-Wishart prior rho ~ IW(df, Sigma), the
conditional posterior given x and kappa is:
\deqn{rho | x, kappa \sim IW(df + N, Sigma + x^T Q x)}
where Q is the marginal precision matrix for one component.
}
\seealso{
\code{\link[=gibbs_kappa_rho]{gibbs_kappa_rho()}}
}
