% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated_landcover}
\alias{simulated_landcover}
\title{Simulated Land Cover Data}
\format{
A list with the following components:
\describe{
\item{y}{Matrix (400 x 3) of observed compositional data. Columns are
Forest, Open, and Other land cover types.}
\item{z_true}{Matrix (400 x 3) of true compositional means.}
\item{x_true}{Matrix (400 x 2) of true latent field in ALR space.}
\item{G}{Sparse precision matrix (400 x 400) for the GMRF.}
\item{coords}{Matrix (400 x 2) of grid coordinates.}
\item{params}{List of true parameter values:
\describe{
\item{kappa}{Spatial range parameter (0.4)}
\item{alpha}{Dirichlet precision (80)}
\item{rho}{2x2 component covariance matrix}
}
}
}
}
\source{
Generated with \code{simulate_landcover(n_side = 20, pattern = "gradient", kappa_true = 0.4, alpha_true = 80, seed = 2024)}
}
\usage{
simulated_landcover
}
\description{
A simulated dataset of spatial compositional land cover data on a 20x20 grid.
}
\details{
The data were simulated using \code{simulate_landcover()} with:
\itemize{
\item Grid size: 20 x 20 = 400 cells
\item Pattern: North-south gradient (more forest in north)
\item Dirichlet precision: 80 (moderate observation noise)
\item Spatial range: kappa = 0.4
\item Cross-component correlation: -0.3 between forest and open land
}
}
\examples{
data(simulated_landcover)

# View dimensions
dim(simulated_landcover$y)

# Plot forest cover
plot_field(simulated_landcover$y[, 1], n_side = 20, main = "Forest")

# True parameters
simulated_landcover$params

}
\keyword{datasets}
