% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatiocompo-package.R
\docType{package}
\name{spatiocompo-package}
\alias{spatiocompo-package}
\alias{spatiocompo}
\title{spatiocompo: Bayesian Spatial Compositional Data Modeling}
\description{
Bayesian spatial compositional data modeling using Gaussian Markov
Random Fields (GMRF) and Dirichlet likelihood. The package implements
Metropolis-Adjusted Langevin Algorithm (MALA) for the latent field and
precision parameter, combined with Gibbs sampling for spatial range and
covariance parameters.
}
\details{
The main function is \code{\link[=mcmc_sampling]{mcmc_sampling()}} which performs posterior inference
for spatial compositional data. The model assumes:

\itemize{
\item Observations \eqn{y_i} follow a Dirichlet distribution with
concentration parameter \eqn{\alpha \cdot z_i}
\item The compositional mean \eqn{z_i} is the inverse ALR transform
of a latent Gaussian field \eqn{x_i}
\item The latent field follows a GMRF prior with Matern-like covariance
\item Covariates can be included through a linear predictor
}

For simulated data examples, see \code{\link[=simulate_compo_data]{simulate_compo_data()}} and
\code{\link[=simulate_landcover]{simulate_landcover()}}.
}
\section{Key Functions}{

\describe{
\item{Compositional transforms}{\code{\link[=alr]{alr()}}, \code{\link[=invalr]{invalr()}}, \code{\link[=dalr]{dalr()}}}
\item{Precision matrices}{\code{\link[=create_Q]{create_Q()}}, \code{\link[=Q_rhoxQ]{Q_rhoxQ()}}, \code{\link[=log_det_Q]{log_det_Q()}}}
\item{MCMC sampling}{\code{\link[=mcmc_sampling]{mcmc_sampling()}}, \code{\link[=mala_step]{mala_step()}}, \code{\link[=gibbs_kappa_rho]{gibbs_kappa_rho()}}}
\item{Simulation}{\code{\link[=simulate_compo_data]{simulate_compo_data()}}, \code{\link[=simulate_landcover]{simulate_landcover()}}}
\item{Visualization}{\code{\link[=plot_results]{plot_results()}}, \code{\link[=plot_trace]{plot_trace()}}}
}
}

\references{
Pirzamanbein, B., Lindstrom, J., Poska, A., Gaillard, M.-J. (2018).
Modelling spatial compositional data: Reconstructions of past land
cover and uncertainties. \emph{Spatial Statistics} 24: 14--31.
\doi{10.1016/j.spasta.2018.03.005}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BehnazP/spatiocompo-r}
  \item Report bugs at \url{https://github.com/BehnazP/spatiocompo-r/issues}
}

}
\author{
Behnaz Pirzamanbin, Johan Lindstrom
}
